requirements.txt
python-telegram-bot==20.7
flask==3.0.2
python-dotenv==1.0.1
